refactors types and models into own files
this commit moves types and models into own files. this hopefully allows to find places to change easier. This commit also starts changing of name from Incident to Attack and from Action to Effect
