+++ game_lib/src/cards/properties/mod.rs
@@ -1,1 +1,8 @@
-pub mod card_content;
+pub mod title;
+pub mod description;
+pub mod effect_description;
+pub mod target;
+pub mod duration;
+pub mod effect;
+pub mod fix_cost;
+pub mod fix_modifier;
