+++ game_lib/src/cards/serialization/mod.rs
@@ -1,1 +1,1 @@
-pub(crate) mod serialization;
+pub(crate) mod helper;
