+++ game_lib/src/cards/types/mod.rs
@@ -1,1 +1,5 @@
 pub mod card_model;
+pub mod oopsie;
+pub mod event;
+pub mod attack;
+pub mod lucky;
