--- game_lib/src/cards/world/game_model.rs
@@ -1,14 +1,0 @@
-#[derive(Clone, Debug)]
-pub struct Resources(pub(crate) usize);
-
-impl Resources {
-    pub fn new(value: usize) -> Self {
-        Resources(value)
-    }
-}
-
-impl Default for Resources {
-    fn default() -> Self {
-        Resources(0)
-    }
-}
